Split stuff in files
