+++ lib/GuacCommon.d.ts
@@ -1,0 +1,12 @@
+// TODO: BigInt?
+export type TypedArray = Uint32Array |
+  Uint16Array |
+  Uint8Array |
+  Uint8ClampedArray |
+  Int32Array |
+  Int16Array |
+  Int8Array |
+  Float64Array |
+  Float32Array;
+
+export type Mimetype = string;
